Move some middleware from excel to sqlController
Some of the middleware in `excelController` should
live in `sqlController` because they are pertain to
generating the entity relationship anyway.

Along the way, fix a small bug where associate
tables are not generated properly
